zep 0.2

+++ src/lib/packages/install.zig
@@ -12,25 +12,30 @@ const UtilsInjector = Utils.UtilsInjector;
 const UtilsPackage = Utils.UtilsPackage;
 const UtilsPrinter = Utils.UtilsPrinter;
 
-const CachePackage =
-    @import("lib/cachePackage.zig");
-const DownloadPackage =
-    @import("lib/downloadPackage.zig");
-const Init =
-    @import("init.zig");
+const CachePackage = @import("lib/cachePackage.zig");
+const DownloadPackage = @import("lib/downloadPackage.zig");
+const Init = @import("init.zig");
+
+/// Check if an array of strings, contains a specific
+/// string
+fn stringInArray(haystack: [][]const u8, needle: []const u8) bool {
+    for (haystack) |h| {
+        if (std.mem.eql(u8, h, needle)) return true;
+    }
+    return false;
+}
 
 pub const Installer = struct {
     allocator: std.mem.Allocator,
-
     json: UtilsJson.Json,
     package: UtilsPackage.Package,
     downloader: DownloadPackage.Downloader,
     cacher: CachePackage.Cacher,
-
     printer: *UtilsPrinter.Printer,
 
     pub fn init(allocator: std.mem.Allocator, printer: *UtilsPrinter.Printer, packageName: ?[]const u8) anyerror!Installer {
         var json = try UtilsJson.Json.init(allocator);
+
         if (packageName == null) {
             try printer.append("Installing all packages...\n");
             try installAll(allocator, &json, printer);
@@ -38,77 +43,71 @@ pub const Installer = struct {
             return error.NoPackage;
         }
 
-        const package = try UtilsPackage.Package.init(allocator, packageName.?, printer);
-        if (package == null) {
+        const package = try UtilsPackage.Package.init(allocator, packageName.?, printer) orelse {
             std.process.exit(0);
             return error.NoPackage;
-        }
+        };
 
-        const cacher = try CachePackage.Cacher.init(allocator, package.?, printer);
-        const downloader = try DownloadPackage.Downloader.init(allocator, package.?, cacher, printer);
+        const cacher = try CachePackage.Cacher.init(allocator, package, printer);
+        const downloader = try DownloadPackage.Downloader.init(allocator, package, cacher, printer);
 
-        return Installer{ .allocator = allocator, .package = package.?, .downloader = downloader, .cacher = cacher, .json = json, .printer = printer };
+        return Installer{
+            .allocator = allocator,
+            .package = package,
+            .downloader = downloader,
+            .cacher = cacher,
+            .json = json,
+            .printer = printer,
+        };
+    }
+
+    pub fn deinit(self: *Installer) void {
+        self.cacher.deinit();
+        self.downloader.deinit();
+        self.package.deinit();
     }
 
     pub fn install(self: *Installer) !void {
-        var package = self.package;
-        if (Locales.VERBOSITY_MODE >= 1) {
-            try self.printer.append("Downloading Package...\n");
-        }
+        var pkg = self.package;
 
+        if (Locales.VERBOSITY_MODE >= 1) try self.printer.append("Downloading Package...\n");
         try self.downloader.downloadPackage();
-        if (Locales.VERBOSITY_MODE >= 1)
-            try self.printer.append("\nChecking fingerprint...\n");
 
-        const fingerprint = try package.checkFingerprint();
-        if (Locales.VERBOSITY_MODE >= 1 and fingerprint)
-            try self.printer.append("FINGERPRINT IDENTICAL!\n");
+        if (Locales.VERBOSITY_MODE >= 1) try self.printer.append("\nChecking fingerprint...\n");
+        if (try pkg.checkFingerprint() and Locales.VERBOSITY_MODE >= 1) try self.printer.append("FINGERPRINT IDENTICAL!\n");
 
-        if (Locales.VERBOSITY_MODE >= 1)
-            try self.printer.append("\nChecking Caching...\n");
+        if (Locales.VERBOSITY_MODE >= 1) try self.printer.append("\nChecking Caching...\n");
+        if (!(try self.cacher.isPackageCached())) {
+            if (Locales.VERBOSITY_MODE >= 1) try self.printer.append("\nCaching...\n");
 
-        const isCached = try self.cacher.isPackageCached();
-        if (isCached) {
-            if (Locales.VERBOSITY_MODE >= 1) {
-                try self.printer.append("PACKAGE ALREADY CACHED! SKIPPING CACHING!\n\n");
-            }
-        } else {
-            if (Locales.VERBOSITY_MODE >= 1)
-                try self.printer.append("\nCaching...\n");
-
-            const cached = try self.cacher.cachePackage();
-            if (cached and Locales.VERBOSITY_MODE >= 1)
-                try self.printer.append("PACKAGE CACHED!\n\n");
-        }
+            try self.cacher.cachePackage();
+            if (Locales.VERBOSITY_MODE >= 1) try self.printer.append("PACKAGE CACHED!\n\n");
+        } else if (Locales.VERBOSITY_MODE >= 1) try self.printer.append("PACKAGE ALREADY CACHED! SKIPPING CACHING!\n\n");
 
         try self.addPackageToJson();
 
         if (Locales.VERBOSITY_MODE >= 1) {
-            const success = try std.fmt.allocPrint(self.allocator, "Successfully installed - {s}\n\n", .{package.packageName});
+            const success = try std.fmt.allocPrint(self.allocator, "Successfully installed - {s}\n\n", .{pkg.packageName});
             try self.printer.append(success);
         }
 
-        var injector = UtilsInjector.Injector.init(self.allocator, package.packageName, self.printer);
+        var injector = UtilsInjector.Injector.init(self.allocator, pkg.packageName, self.printer);
         try injector.initInjector();
 
-        // create a symbolic link
-        const targetPath = try std.fmt.allocPrint(self.allocator, "{s}/{s}", .{ Constants.ROOT_ZEP_PKG_FOLDER, package.packageName });
+        // symbolic link
+        const targetPath = try std.fmt.allocPrint(self.allocator, "{s}/{s}", .{ Constants.ROOT_ZEP_PKG_FOLDER, pkg.packageName });
         defer self.allocator.free(targetPath);
 
-        const linkPath = try std.fmt.allocPrint(self.allocator, "{s}/{s}", .{ Constants.ZEP_FOLDER, package.packageName });
+        const linkPath = try std.fmt.allocPrint(self.allocator, "{s}/{s}", .{ Constants.ZEP_FOLDER, pkg.packageName });
         defer self.allocator.free(linkPath);
 
         const cwd = try std.fs.cwd().realpathAlloc(self.allocator, ".");
         defer self.allocator.free(cwd);
+
         const absLinkedPath = try std.fs.path.resolve(self.allocator, &[_][]const u8{ cwd, linkPath });
         defer self.allocator.free(absLinkedPath);
 
-        // Delete old link if it exists
-        if (try UtilsFs.checkDirExists(linkPath)) {
-            try std.fs.cwd().deleteDir(linkPath);
-        }
-
-        // Create the symlink
+        if (try UtilsFs.checkDirExists(linkPath)) try std.fs.cwd().deleteDir(linkPath);
         try std.fs.cwd().symLink(targetPath, linkPath, .{ .is_directory = true });
 
         try self.addPathToManifest(absLinkedPath);
@@ -129,20 +128,14 @@ pub const Installer = struct {
         defer pkgJson.?.deinit();
         defer lckJson.?.deinit();
 
-        var packageJson = pkgJson.?.value;
-        var lockJson = lckJson.?.value;
-
-        var package = self.package;
-        try package.pkgAppendPackage(&packageJson);
-        try package.lockAppendPackage(&lockJson);
+        try self.package.pkgAppendPackage(&pkgJson.?.value);
+        try self.package.lockAppendPackage(&lckJson.?.value);
     }
 
     pub fn addPathToManifest(self: *Installer, linkedPath: []const u8) !void {
         const pkgManifest = try self.json.parsePkgManifest();
         var pkgVal: Structs.PkgsManifest = Structs.PkgsManifest{ .packages = &[_]Structs.PkgManifest{} };
-        defer {
-            if (pkgManifest) |pkg| pkg.deinit();
-        }
+        defer if (pkgManifest) |pkg| pkg.deinit();
         if (pkgManifest) |pkg| pkgVal = pkg.value;
 
         var list = std.ArrayList(Structs.PkgManifest).init(self.allocator);
@@ -153,24 +146,14 @@ pub const Installer = struct {
 
         for (pkgVal.packages) |p| {
             if (std.mem.eql(u8, p.name, self.package.packageName)) {
-                for (p.paths) |path| {
-                    try listPath.append(path);
-                }
+                for (p.paths) |path| try listPath.append(path);
                 continue;
             }
             try list.append(p);
         }
 
-        var isIn = false;
-        for (listPath.items) |path| {
-            if (std.mem.eql(u8, path, linkedPath)) {
-                isIn = true;
-                break;
-            }
-        }
-        if (!isIn) {
-            try listPath.append(linkedPath);
-        }
+        if (!stringInArray(listPath.items, linkedPath)) try listPath.append(linkedPath);
+        // if (!std.mem.indexOf([]u8, listPath.items, linkedPath)) try listPath.append(linkedPath);
         try list.append(Structs.PkgManifest{ .name = self.package.packageName, .paths = listPath.items });
 
         pkgVal.packages = list.items;
@@ -185,14 +168,12 @@ fn installAll(allocator: std.mem.Allocator, json: *UtilsJson.Json, printer: *Uti
     if (pkgJsonOpt == null) {
         var initter = try Init.Init.init(allocator);
         try initter.commitInit();
-
         try printer.append("zep.json not initialized. Initializing...\n");
         try printer.append("Nothing to install...\n");
         return;
     }
 
     Locales.VERBOSITY_MODE = 0;
-
     const pkgJson = pkgJsonOpt.?.value;
     defer pkgJsonOpt.?.deinit();
